test: share task_name_status and user_with_email scenarios
The `Task { id, name, status }` model with a unit `Status { Pending, Active,
Done }` enum was duplicated across four filter tests in three files; extract
it into a task_name_status scenario. The minimal `User { id, email: Email }`
newtype model in basic_newtype_embed and newtype_column_name moves to a
user_with_email scenario.
