+++ crates/toasty-driver-integration-suite/src/scenarios.rs
@@ -21,6 +21,7 @@ pub mod in_list_item;
 pub mod person_contact_deferred_metadata;
 pub mod tagged_item;
 pub mod task_bug_priority;
+pub mod task_name_status;
 pub mod task_with_status;
 pub mod task_with_string_status;
 pub mod two_models;
@@ -40,6 +41,7 @@ pub mod user_unique_email_with_name;
 pub mod user_with_active;
 pub mod user_with_address;
 pub mod user_with_age;
+pub mod user_with_email;
 pub mod user_with_status;
 pub mod user_with_zip_address;
 pub mod versioned_item;
