+++ crates/toasty-driver-integration-suite/src/scenarios/task_name_status.rs
@@ -1,0 +1,31 @@
+use crate::prelude::*;
+
+scenario! {
+    #![id(ID)]
+
+    #[derive(Debug, toasty::Model)]
+    #[allow(dead_code)]
+    struct Task {
+        #[key]
+        #[auto]
+        id: ID,
+
+        name: String,
+
+        status: Status,
+    }
+
+    #[derive(Debug, PartialEq, toasty::Embed)]
+    enum Status {
+        #[column(variant = 1)]
+        Pending,
+        #[column(variant = 2)]
+        Active,
+        #[column(variant = 3)]
+        Done,
+    }
+
+    async fn setup(test: &mut Test) -> toasty::Db {
+        test.setup_db(models!(Task)).await
+    }
+}
