+++ crates/toasty-driver-integration-suite/src/scenarios/user_with_email.rs
@@ -1,0 +1,19 @@
+use crate::prelude::*;
+
+scenario! {
+    #[derive(Debug, toasty::Model)]
+    #[allow(dead_code)]
+    struct User {
+        #[key]
+        id: String,
+
+        email: Email,
+    }
+
+    #[derive(Debug, toasty::Embed)]
+    struct Email(String);
+
+    async fn setup(test: &mut Test) -> toasty::Db {
+        test.setup_db(models!(User)).await
+    }
+}
